// hr-frontend/src/lib/api/generated/model/contractTemplateSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ContractTemplateSummary {
  contractTemplateId?: number;
  contractTypeId?: number;
  templateCode?: string;
  versionNumber?: number;
  templateNameEn?: string;
  templateNameAr?: string;
  effectiveFrom?: string;
  effectiveTo?: string;
  active?: boolean;
  fileUploaded?: boolean;
}